backend/requirements.txt
fastapi
uvicorn
websockets
langgraph
langchain-openai
langchain-core
langchain-community
python-dotenv
colorama
uvicorn